refactor: parsefilters to util file
