+++ utils/parseFilters.js
@@ -1,0 +1,18 @@
+function parseFilters(filtersString) {
+  if (!filtersString) return {}
+
+  const filters = {}
+  const conditions = filtersString.split(' AND ')
+
+  conditions.forEach((condition) => {
+    const [key, value] = condition.split(':(')
+    const cleanedKey = key.trim()
+    const values = value.replace(')', '').split(' OR ').map(v => v.trim())
+
+    filters[cleanedKey] = values
+  })
+
+  return filters
+}
+
+export default parseFilters
